requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.8
scikit-learn>=1.3
pillow>=10.0
opencv-python>=4.8
tqdm>=4.66
pyyaml>=6.0
rich>=13.7
torch>=2.1
torchvision>=0.16
timm>=0.9
